pull main into my branch

+++ src/main/java/org/arbeitspferde/groningen/extractor/Extractor.java
@@ -16,6 +16,7 @@
 package org.arbeitspferde.groningen.extractor;
 
 import com.google.common.base.Optional;
+
 import org.arbeitspferde.groningen.config.GroningenConfig;
 import org.arbeitspferde.groningen.experimentdb.SubjectStateBridge;
 import org.arbeitspferde.groningen.subject.Subject;
@@ -24,8 +25,8 @@ import org.arbeitspferde.groningen.utility.Metric;
 import org.arbeitspferde.groningen.utility.MetricExporter;
 
 import javax.annotation.Nullable;
+
 import java.io.BufferedReader;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.util.StringTokenizer;
@@ -152,7 +153,7 @@ public class Extractor implements Runnable {
               double pauseTimeSecs;
 
               try {
-                pauseTimeSecs = Double.valueOf(token.trim()).doubleValue();
+                pauseTimeSecs = Double.valueOf(token.trim());
               } catch (final Exception e) {
                 log.log(Level.WARNING,
                     String.format("Unable to parse pause time '%s'. Defaulting to 0.0", token), e);
